dev/backend/app/tools/mcp_server/mcp_server_web_search/pyproject.toml
[project]
name = "mcp-server-web-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.3"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "duckduckgo-search>=8.0.3",
    "requests>=2.32.4",
    "fastmcp>=2.8.0",
    "mcp-server-fetch>=2025.1.17",
]
